🔧 REFACTOR: Reorganize calling functions into separate files and add PR checks workflow
